optimize: make shadow tls a general transport wrapper, like tls
