--- clash_lib/src/proxy/shadowsocks/shadow_tls/mod.rs
@@ -1,6 +1,0 @@
-mod connector;
-mod stream;
-mod utils;
-
-pub use connector::Connector;
-use utils::prelude;
